test(pie-docs): DSW-3266 moved playwright configs to pie-docs
